backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
celery==5.3.6
redis==5.0.4
sqlalchemy==2.0.30
aiofiles==23.2.1
python-multipart==0.0.9
anthropic==0.25.0
httpx==0.27.0
python-dotenv==1.0.1
ffmpeg-python==0.2.0
Pillow==10.3.0
